Deduplicate tons of code
