+++ pkg/reconciler/notification_store_test.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package reconciler
+package reconciler_test
 
 import (
 	"testing"
@@ -23,70 +23,42 @@ import (
 	"go.uber.org/atomic"
 	"k8s.io/apimachinery/pkg/types"
 	control "knative.dev/control-protocol/pkg"
+	"knative.dev/control-protocol/pkg/reconciler"
+	"knative.dev/control-protocol/pkg/test"
 )
 
-func setupNotificationStoreTest(t *testing.T) (control.Service, *atomic.Int32, *NotificationStore, types.NamespacedName, string) {
-	expectedNamespacedName := types.NamespacedName{Namespace: "hello", Name: "world"}
-	expectedPodIp := "127.0.0.1"
+func TestNotificationStore_Integration_StoreMessages(t *testing.T) {
+	controlPlane, enqueueKeyInvoked, notificationsStore, expectedNamespacedName, expectedPodIp := setupNotificationStoreIntegrationTest(t, test.MockMessageMerger)
 
-	controlPlane, dataPlane := setupConnection(t)
-
-	enqueueKeyInvoked := atomic.NewInt32(0)
-
-	notificationsStore := NewNotificationStore(func(name types.NamespacedName) {
-		require.Equal(t, expectedNamespacedName, name)
-		enqueueKeyInvoked.Inc()
-	}, parseMockMessage)
-
-	dataPlane.MessageHandler(notificationsStore.ControlMessageHandler(expectedNamespacedName, expectedPodIp, mockValueMerger))
-
-	return controlPlane, enqueueKeyInvoked, notificationsStore, expectedNamespacedName, expectedPodIp
-}
-
-func TestNotificationStore_StoreMessages(t *testing.T) {
-	controlPlane, enqueueKeyInvoked, notificationsStore, expectedNamespacedName, expectedPodIp := setupNotificationStoreTest(t)
-
-	require.NoError(t, controlPlane.SendAndWaitForAck(1, mockMessage("Funky!")))
-	require.NoError(t, controlPlane.SendAndWaitForAck(1, mockMessage("More Funky!")))
+	require.NoError(t, controlPlane.SendAndWaitForAck(1, test.MockPayload("Funky!")))
+	require.NoError(t, controlPlane.SendAndWaitForAck(1, test.MockPayload("More Funky!")))
 
 	require.Equal(t, int32(2), enqueueKeyInvoked.Load())
 
 	stored, ok := notificationsStore.GetPodNotification(expectedNamespacedName, expectedPodIp)
 	require.True(t, ok)
 
-	value := stored.(*mockMessage)
+	value := stored.(*test.MockPayload)
 	require.Equal(t, "Funky!More Funky!", string(*value))
 }
 
-func TestNotificationStore_DontReconcileTwice(t *testing.T) {
-	expectedNamespacedName := types.NamespacedName{Namespace: "hello", Name: "world"}
-	expectedPodIp := "127.0.0.1"
+func TestNotificationStore_Integration_DontReconcileTwice(t *testing.T) {
+	controlPlane, enqueueKeyInvoked, notificationsStore, expectedNamespacedName, expectedPodIp := setupNotificationStoreIntegrationTest(t, reconciler.PassNewValue)
 
-	controlPlane, dataPlane := setupConnection(t)
-
-	enqueueKeyInvoked := atomic.NewInt32(0)
-
-	notificationsStore := NewNotificationStore(func(name types.NamespacedName) {
-		require.Equal(t, expectedNamespacedName, name)
-		enqueueKeyInvoked.Inc()
-	}, parseMockMessage)
-
-	dataPlane.MessageHandler(notificationsStore.ControlMessageHandler(expectedNamespacedName, expectedPodIp, PassNewValue))
-
-	require.NoError(t, controlPlane.SendAndWaitForAck(1, mockMessage("Funky!")))
-	require.NoError(t, controlPlane.SendAndWaitForAck(1, mockMessage("Funky!")))
+	require.NoError(t, controlPlane.SendAndWaitForAck(1, test.MockPayload("Funky!")))
+	require.NoError(t, controlPlane.SendAndWaitForAck(1, test.MockPayload("Funky!")))
 
 	require.Equal(t, int32(1), enqueueKeyInvoked.Load())
 
 	stored, ok := notificationsStore.GetPodNotification(expectedNamespacedName, expectedPodIp)
 	require.True(t, ok)
 
-	value := stored.(*mockMessage)
+	value := stored.(*test.MockPayload)
 	require.Equal(t, "Funky!", string(*value))
 }
 
-func TestNotificationStore_GetMessages(t *testing.T) {
-	controlPlane, enqueueKeyInvoked, notificationsStore, expectedNamespacedName, expectedPodIp := setupNotificationStoreTest(t)
+func TestNotificationStore_Integration_GetMessages(t *testing.T) {
+	controlPlane, enqueueKeyInvoked, notificationsStore, expectedNamespacedName, expectedPodIp := setupNotificationStoreIntegrationTest(t, test.MockMessageMerger)
 
 	stored, ok := notificationsStore.GetPodNotification(expectedNamespacedName, expectedPodIp)
 	require.False(t, ok)
@@ -96,7 +68,7 @@ func TestNotificationStore_GetMessages(t *testing.T) {
 	require.False(t, ok)
 	require.Nil(t, stored)
 
-	require.NoError(t, controlPlane.SendAndWaitForAck(1, mockMessage("Funky!")))
+	require.NoError(t, controlPlane.SendAndWaitForAck(1, test.MockPayload("Funky!")))
 
 	require.Equal(t, int32(1), enqueueKeyInvoked.Load())
 
@@ -106,27 +78,27 @@ func TestNotificationStore_GetMessages(t *testing.T) {
 
 	stored, ok = notificationsStore.GetPodNotification(expectedNamespacedName, expectedPodIp)
 	require.True(t, ok)
-	value := stored.(*mockMessage)
+	value := stored.(*test.MockPayload)
 	require.Equal(t, "Funky!", string(*value))
 }
 
-func TestNotificationStore_ClearMessagesWithoutAny(t *testing.T) {
-	_, _, notificationsStore, expectedNamespacedName, expectedPodIp := setupNotificationStoreTest(t)
+func TestNotificationStore_Integration_ClearMessagesWithoutAny(t *testing.T) {
+	_, _, notificationsStore, expectedNamespacedName, expectedPodIp := setupNotificationStoreIntegrationTest(t, test.MockMessageMerger)
 
 	notificationsStore.CleanPodNotification(expectedNamespacedName, expectedPodIp)
 	notificationsStore.CleanPodsNotifications(expectedNamespacedName)
 }
 
-func TestNotificationStore_ClearMessages(t *testing.T) {
-	controlPlane, enqueueKeyInvoked, notificationsStore, expectedNamespacedName, expectedPodIp := setupNotificationStoreTest(t)
+func TestNotificationStore_Integration_ClearMessages(t *testing.T) {
+	controlPlane, enqueueKeyInvoked, notificationsStore, expectedNamespacedName, expectedPodIp := setupNotificationStoreIntegrationTest(t, test.MockMessageMerger)
 
-	require.NoError(t, controlPlane.SendAndWaitForAck(1, mockMessage("Funky!")))
+	require.NoError(t, controlPlane.SendAndWaitForAck(1, test.MockPayload("Funky!")))
 
 	require.Equal(t, int32(1), enqueueKeyInvoked.Load())
 
 	stored, ok := notificationsStore.GetPodNotification(expectedNamespacedName, expectedPodIp)
 	require.True(t, ok)
-	value := stored.(*mockMessage)
+	value := stored.(*test.MockPayload)
 	require.Equal(t, "Funky!", string(*value))
 
 	notificationsStore.CleanPodNotification(expectedNamespacedName, expectedPodIp)
@@ -134,4 +106,22 @@ func TestNotificationStore_ClearMessages(t *testing.T) {
 	stored, ok = notificationsStore.GetPodsNotifications(expectedNamespacedName)
 	require.False(t, ok)
 	require.Nil(t, stored)
+}
+
+func setupNotificationStoreIntegrationTest(t *testing.T, merger reconciler.ValueMerger) (control.Service, *atomic.Int32, *reconciler.NotificationStore, types.NamespacedName, string) {
+	expectedNamespacedName := types.NamespacedName{Namespace: "hello", Name: "world"}
+	expectedPodIp := "127.0.0.1"
+
+	_, dataPlane, _, controlPlane := test.MustSetupInsecureControlPair(t)
+
+	enqueueKeyInvoked := atomic.NewInt32(0)
+
+	notificationsStore := reconciler.NewNotificationStore(func(name types.NamespacedName) {
+		require.Equal(t, expectedNamespacedName, name)
+		enqueueKeyInvoked.Inc()
+	}, test.ParseMockMessage)
+
+	dataPlane.MessageHandler(notificationsStore.ControlMessageHandler(expectedNamespacedName, expectedPodIp, merger))
+
+	return controlPlane, enqueueKeyInvoked, notificationsStore, expectedNamespacedName, expectedPodIp
 }
