+++ pkg/service/message_router_test.go
@@ -1,4 +1,20 @@
-package service
+/*
+Copyright 2021 The Knative Authors
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package service_test
 
 import (
 	"context"
@@ -9,6 +25,7 @@ import (
 	"go.uber.org/atomic"
 
 	ctrl "knative.dev/control-protocol/pkg"
+	"knative.dev/control-protocol/pkg/service"
 	"knative.dev/control-protocol/pkg/test"
 )
 
@@ -17,7 +34,7 @@ func TestMessageRouter_HandleServiceMessage_Matching(t *testing.T) {
 
 	counter := atomic.NewInt32(0)
 
-	svc.MessageHandler(MessageRouter{
+	svc.MessageHandler(service.MessageRouter{
 		1: ctrl.MessageHandlerFunc(func(ctx context.Context, message ctrl.ServiceMessage) {
 			counter.Inc()
 		}),
@@ -33,7 +50,7 @@ func TestMessageRouter_HandleServiceMessage_NotMatching(t *testing.T) {
 
 	counter := atomic.NewInt32(0)
 
-	svc.MessageHandler(MessageRouter{
+	svc.MessageHandler(service.MessageRouter{
 		1: ctrl.MessageHandlerFunc(func(ctx context.Context, message ctrl.ServiceMessage) {
 			counter.Inc()
 		}),
@@ -42,4 +59,60 @@ func TestMessageRouter_HandleServiceMessage_NotMatching(t *testing.T) {
 	inboundMsg := ctrl.NewInboundMessage(ctrl.ActualProtocolVersion, 0, 10, uuid.New(), nil)
 	require.True(t, svc.InvokeMessageHandler(context.TODO(), &inboundMsg))
 	require.Equal(t, int32(0), counter.Load())
+}
+
+func TestMessageRouterIntegration(t *testing.T) {
+	_, dataPlane, _, controlPlane := test.MustSetupInsecureControlPair(t)
+
+	opcode1Count := atomic.NewInt32(0)
+	opcode2Count := atomic.NewInt32(0)
+
+	dataPlane.MessageHandler(service.MessageRouter{
+		1: ctrl.MessageHandlerFunc(func(ctx context.Context, message ctrl.ServiceMessage) {
+			require.Equal(t, uint8(1), message.Headers().OpCode())
+			require.Equal(t, "Funky!", string(message.Payload()))
+			message.Ack()
+			opcode1Count.Inc()
+		}),
+		2: ctrl.MessageHandlerFunc(func(ctx context.Context, message ctrl.ServiceMessage) {
+			require.Equal(t, uint8(2), message.Headers().OpCode())
+			require.Equal(t, "Funky!", string(message.Payload()))
+			message.Ack()
+			opcode2Count.Inc()
+		}),
+	})
+
+	for i := 0; i < 10; i++ {
+		require.NoError(t, controlPlane.SendAndWaitForAck(ctrl.OpCode((i%2)+1), test.MockPayload("Funky!")))
+	}
+
+	require.Equal(t, int32(5), opcode1Count.Load())
+	require.Equal(t, int32(5), opcode2Count.Load())
+}
+
+func TestMessageRouterIntegration_MessageNotMatchingAck(t *testing.T) {
+	_, dataPlane, _, controlPlane := test.MustSetupInsecureControlPair(t)
+
+	opcode1Count := atomic.NewInt32(0)
+	opcode2Count := atomic.NewInt32(0)
+
+	dataPlane.MessageHandler(service.MessageRouter{
+		1: ctrl.MessageHandlerFunc(func(ctx context.Context, message ctrl.ServiceMessage) {
+			require.Equal(t, uint8(1), message.Headers().OpCode())
+			require.Equal(t, "Funky!", string(message.Payload()))
+			message.Ack()
+			opcode1Count.Inc()
+		}),
+		2: ctrl.MessageHandlerFunc(func(ctx context.Context, message ctrl.ServiceMessage) {
+			require.Equal(t, uint8(2), message.Headers().OpCode())
+			require.Equal(t, "Funky!", string(message.Payload()))
+			message.Ack()
+			opcode2Count.Inc()
+		}),
+	})
+
+	require.NoError(t, controlPlane.SendAndWaitForAck(10, test.MockPayload("Funky!")))
+
+	require.Equal(t, int32(0), opcode1Count.Load())
+	require.Equal(t, int32(0), opcode2Count.Load())
 }
