+++ pkg/service/service_caching_wrapper_test.go
@@ -1,4 +1,20 @@
-package service
+/*
+Copyright 2021 The Knative Authors
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package service_test
 
 import (
 	"context"
@@ -9,6 +25,7 @@ import (
 	"github.com/stretchr/testify/require"
 
 	control "knative.dev/control-protocol/pkg"
+	"knative.dev/control-protocol/pkg/service"
 	"knative.dev/control-protocol/pkg/test"
 )
 
@@ -18,7 +35,7 @@ func TestCachingService(t *testing.T) {
 	ctx, cancelFn := context.WithCancel(context.TODO())
 	t.Cleanup(cancelFn)
 
-	svc := WithCachingService(ctx)(NewService(ctx, mockConnection))
+	svc := service.WithCachingService(ctx)(service.NewService(ctx, mockConnection))
 
 	var wg sync.WaitGroup
 	wg.Add(1)
