+++ pkg/test/payload_marshaller_mock.go
@@ -1,3 +1,19 @@
+/*
+Copyright 2021 The Knative Authors
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
 package test
 
 type MockPayload string
@@ -11,4 +27,20 @@ func (m MockPayload) MarshalBinary() (data []byte, err error) {
 func (m *MockPayload) UnmarshalBinary(data []byte) error {
 	*m = MockPayload(data)
 	return nil
+}
+
+func ParseMockMessage(bytes []byte) (interface{}, error) {
+	var msg MockPayload
+	err := (&msg).UnmarshalBinary(bytes)
+	if err != nil {
+		return nil, err
+	}
+	return &msg, nil
+}
+
+func MockMessageMerger(old interface{}, new interface{}) interface{} {
+	oldMsg := old.(*MockPayload)
+	newMsg := new.(*MockPayload)
+	merged := MockPayload(string(*oldMsg) + string(*newMsg))
+	return &merged
 }
